requirements.txt
Django==1.8.3
#django-allauth==0.21.0
#https://github.com/pennersr/django-allauth/issues/1076
django-allauth==0.23.0
django-rest-auth==0.3.4
djangorestframework==3.0.5
Pillow==2.8.1
#PyMySQL==0.6.6
mysqlclient==1.3.6
django-material==0.4.1
django-haystack==2.4.0
#celery-haystack==0.9
shortuuid==0.4.2

#from link
#django-haystack-rqueue==0.2.1a0
django-taggit==0.15.0
Whoosh==2.7.0
#django-comments-threaded==1.0.0
gunicorn==19.3.0
django-redis-sessions==0.5.0
beautifulsoup4==4.4.0
lxml==3.4.4
#Twisted==15.2.1
aiohttp==0.16.6
django-mobile==0.5.0